Attend to reviews on refactor
