+++ crates/typst-svg/src/shape.rs
@@ -1,10 +1,8 @@
-use base64::Engine;
-use ecow::{eco_format, EcoString};
+use ecow::EcoString;
 use ttf_parser::OutlineBuilder;
-use typst::layout::{Abs, Axes, Ratio, Size, Transform};
+use typst::layout::{Abs, Ratio, Size, Transform};
 use typst::visualize::{
-    FixedStroke, Geometry, Image, ImageFormat, LineCap, LineJoin, Paint, Path, PathItem,
-    RasterFormat, RelativeTo, Shape, VectorFormat,
+    FixedStroke, Geometry, LineCap, LineJoin, Paint, Path, PathItem, RelativeTo, Shape,
 };
 
 use crate::paint::ColorEncode;
@@ -151,17 +149,6 @@ impl SVGRenderer {
             );
         }
     }
-
-    /// Render an image element.
-    pub(super) fn render_image(&mut self, image: &Image, size: &Axes<Abs>) {
-        let url = convert_image_to_base64_url(image);
-        self.xml.start_element("image");
-        self.xml.write_attribute("xlink:href", &url);
-        self.xml.write_attribute("width", &size.x.to_pt());
-        self.xml.write_attribute("height", &size.y.to_pt());
-        self.xml.write_attribute("preserveAspectRatio", "none");
-        self.xml.end_element();
-    }
 }
 
 /// Convert a geometry to an SVG path.
@@ -205,25 +192,4 @@ pub fn convert_path(path: &Path) -> EcoString {
         }
     }
     builder.0
-}
-
-/// Encode an image into a data URL. The format of the URL is
-/// `data:image/{format};base64,`.
-#[comemo::memoize]
-pub fn convert_image_to_base64_url(image: &Image) -> EcoString {
-    let format = match image.format() {
-        ImageFormat::Raster(f) => match f {
-            RasterFormat::Png => "png",
-            RasterFormat::Jpg => "jpeg",
-            RasterFormat::Gif => "gif",
-        },
-        ImageFormat::Vector(f) => match f {
-            VectorFormat::Svg => "svg+xml",
-        },
-    };
-
-    let mut url = eco_format!("data:image/{format};base64,");
-    let data = base64::engine::general_purpose::STANDARD.encode(image.data());
-    url.push_str(&data);
-    url
 }
